server/requirements.txt
 Django==2.0.3
 psycopg2
 djangorestframework